+++ src/Bag.test.ts
@@ -7,7 +7,7 @@
 
 /* eslint-disable jest/no-conditional-expect */
 
-import { compare, fromDate } from "@foxglove/rostime";
+import { compare, fromDate } from "@lichtblick/rostime";
 import compress from "compressjs";
 import fs from "fs";
 import lz4 from "lz4js";
@@ -186,7 +186,9 @@ describe("Bag", () => {
     const messages = await fullyReadBag(FILENAME, { topics: ["/turtle1/color_sensor"] });
     const topics = messages.map((msg) => msg.topic);
     expect(topics).toHaveLength(1351);
-    topics.forEach((topic) => expect(topic).toBe("/turtle1/color_sensor"));
+    topics.forEach((topic) => {
+      expect(topic).toBe("/turtle1/color_sensor");
+    });
   });
 
   it("reads messages filtered to multiple topics", async () => {
@@ -194,9 +196,9 @@ describe("Bag", () => {
     const messages = await fullyReadBag(FILENAME, opts);
     const topics = messages.map((msg) => msg.topic);
     expect(topics).toHaveLength(2695);
-    topics.forEach((topic) =>
-      expect(topic === "/turtle1/color_sensor" || topic === "/turtle2/color_sensor").toBe(true),
-    );
+    topics.forEach((topic) => {
+      expect(topic === "/turtle1/color_sensor" || topic === "/turtle2/color_sensor").toBe(true);
+    });
   });
 
   it("reads messages from a shuffled bag", async () => {
@@ -244,7 +246,9 @@ describe("Bag", () => {
       const filename = getFixture("example-bz2");
       const reader = new FileReader(filename);
       const bag = await open(reader);
-      await expect(async () => await bag.readMessages({}, () => {})).rejects.toThrow("compression");
+      await expect(async () => {
+        await bag.readMessages({}, () => {});
+      }).rejects.toThrow("compression");
       await reader.close();
     });
 
@@ -259,7 +263,9 @@ describe("Bag", () => {
       });
       const topics = messages.map((msg) => msg.topic);
       expect(topics).toHaveLength(1351);
-      topics.forEach((topic) => expect(topic).toBe("/turtle1/color_sensor"));
+      topics.forEach((topic) => {
+        expect(topic).toBe("/turtle1/color_sensor");
+      });
     });
 
     it("reads lz4 with supplied decompression callback", async () => {
@@ -271,7 +277,9 @@ describe("Bag", () => {
       });
       const topics = messages.map((msg) => msg.topic);
       expect(topics).toHaveLength(1351);
-      topics.forEach((topic) => expect(topic).toBe("/turtle1/color_sensor"));
+      topics.forEach((topic) => {
+        expect(topic).toBe("/turtle1/color_sensor");
+      });
     });
 
     it("calls decompress with the chunk size", async () => {