+++ CHANGELOG.md
@@ -14,6 +14,6 @@ First release after fork from rosbag.js
 - Conversion to typescript
 - Change APIs to use Uint8Array instead of node.js Buffer
 - Change FileReader and BlobReader to be explicit entry points
-- Remove MessageReader and MessageWriter - these are now in the @foxglove/rosmsg-serialization package
-- Remove parseMessage - this is now in the @foxglove/rosmsg package
+- Remove MessageReader and MessageWriter - these are now in the @lichtblick/rosmsg-serialization package
+- Remove parseMessage - this is now in the @lichtblick/rosmsg package
 - Remove int53 dependency - use BigInt where possible